Hand instantiation
Modified the 'Hand' class to manage the user or dealer's cards, as it previously did, and handle their display, which the 'HandDisplay' class no longer does. The 'Hand' class now instantiates representations of hands using the 'Hand' FXML file to increase their reusability and in preparation for user splitting functionality.